+++ amplipi/streams.py
@@ -362,7 +362,11 @@ class RCA(BaseStream):
       self.disabled = kwargs['disabled']
 
   def info(self) -> models.SourceInfo:
-    src_info = models.SourceInfo(img_url='static/imgs/rca_inputs.svg', name=self.full_name(), state='stopped')
+    src_info = models.SourceInfo(
+      type=self.stream_type,
+      img_url='static/imgs/rca_inputs.svg',
+      name=self.full_name(),
+      state='stopped')
     playing = False
     status_file = f'{utils.get_folder("config")}/srcs/rca_status'
     try:
@@ -505,7 +509,8 @@ class AirPlay(PersistentStream):
     source = models.SourceInfo(
       name=f"Connect to {self.name} on Airplay{'2' if self.ap2 else ''}",
       state=self.state,
-      img_url='static/imgs/shairport.png'
+      img_url='static/imgs/shairport.png',
+      type=self.stream_type
     )
 
     # if stream is airplay2 and other airplay2s exist show error message
@@ -648,7 +653,8 @@ class Spotify(PersistentStream):
     source = models.SourceInfo(
       name=self.full_name(),
       state=self.state,
-      img_url='static/imgs/spotify.png'  # report generic spotify image in place of unspecified album art
+      img_url='static/imgs/spotify.png',  # report generic spotify image in place of unspecified album art
+      type=self.stream_type
     )
     if self.mpris is None:
       return source
@@ -818,7 +824,8 @@ class Pandora(PersistentStream, Browsable):
       name=self.full_name(),
       state=self.state,
       supported_cmds=list(self.supported_cmds.keys()),
-      img_url='static/imgs/pandora.png'
+      img_url='static/imgs/pandora.png',
+      type=self.stream_type
     )
     try:
       with open(loc, 'r', encoding='utf-8') as file:
@@ -1027,7 +1034,12 @@ class DLNA(BaseStream): # TODO: make DLNA a persistent stream to fix the uuid is
     self.dlna_proc = None
 
   def info(self) -> models.SourceInfo:
-    source = models.SourceInfo(name=self.full_name(), state=self.state, img_url='static/imgs/dlna.png')
+    source = models.SourceInfo(
+      name=self.full_name(),
+      state=self.state,
+      img_url='static/imgs/dlna.png',
+      type=self.stream_type
+    )
     try:
 
       data = json.load(open(f'{self._src_config_folder}/meta.json'))
@@ -1131,7 +1143,8 @@ class InternetRadio(BaseStream):
     source = models.SourceInfo(name=self.full_name(),
                                state=self.state,
                                img_url='static/imgs/internet_radio.png',
-                               supported_cmds=self.supported_cmds)
+                               supported_cmds=self.supported_cmds,
+                               type=self.stream_type)
     if self.logo:
       source.img_url = self.logo
     try:
@@ -1196,7 +1209,12 @@ class Plexamp(BaseStream):
     self._disconnect()
 
   def info(self) -> models.SourceInfo:
-    source = models.SourceInfo(name=self.full_name(), state=self.state, img_url='static/imgs/plexamp.png')
+    source = models.SourceInfo(
+      name=self.full_name(),
+      state=self.state,
+      img_url='static/imgs/plexamp.png',
+      type=self.stream_type
+    )
     source.track = "Not currently supported"
     return source
 
@@ -1242,7 +1260,8 @@ class Aux(BaseStream):
   def info(self) -> models.SourceInfo:
     source = models.SourceInfo(name=self.full_name(),
                                img_url='static/imgs/aux_input.svg',
-                               state=self.state)
+                               state=self.state,
+                               type=self.stream_type)
     return source
 
 
@@ -1305,7 +1324,12 @@ class FilePlayer(BaseStream):
     self._disconnect()
 
   def info(self) -> models.SourceInfo:
-    source = models.SourceInfo(name=self.full_name(), state=self.state, img_url='static/imgs/plexamp.png')
+    source = models.SourceInfo(
+      name=self.full_name(),
+      state=self.state,
+      img_url='static/imgs/plexamp.png',
+      type=self.stream_type
+    )
     return source
 
 
@@ -1372,7 +1396,12 @@ class FMRadio(BaseStream):
     loc = f'{src_config_folder}/currentSong'
     if not self.logo:
       self.logo = "static/imgs/fmradio.png"
-    source = models.SourceInfo(name=self.full_name(), state=self.state, img_url=self.logo)
+    source = models.SourceInfo(
+      name=self.full_name(),
+      state=self.state,
+      img_url=self.logo,
+      type=self.stream_type
+    )
     try:
       with open(loc, 'r', encoding='utf-8') as file:
         data = json.loads(file.read())
@@ -1558,7 +1587,8 @@ class LMS(PersistentStream):
       img_url= self.meta.get('image_url', ''),
       track= self.meta.get('track', ''),
       album= self.meta.get('album', ''),
-      artist= self.meta.get('artist', '')
+      artist= self.meta.get('artist', ''),
+      type=self.stream_type
     )
     return source
 
@@ -1639,7 +1669,8 @@ class Bluetooth(BaseStream):
     source = models.SourceInfo(name=self.full_name(),
                                state=self.state,
                                img_url=self.logo,
-                               supported_cmds=self.supported_cmds)
+                               supported_cmds=self.supported_cmds,
+                               type=self.stream_type)
     try:
       with open(loc, 'r') as file:
         data = json.loads(file.read())