+++ web/src/components/StreamBadge/StreamBadge.jsx
@@ -9,9 +9,8 @@ import PropTypes from "prop-types";
 const StreamBadge = ({ sourceId, onClick }) => {
     const info = useStatusStore((s) => s.status.sources[sourceId].info);
     const name = info.name.split(" - ")[0];
-    const type = info.name.split(" - ")[1];
 
-    const icon = getIcon(type);
+    const icon = getIcon(info.type);
 
     return (
         <Chip onClick={onClick} style={{ maxWidth: "60vw" }}>