+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * Add restart stream button to stream player and stream modal
   * Add dynamic scaling to controls on Player page
   * Make update available badge show up on both Settings page and the menu bar
+  * Fix stream icons
 * System
   * Add serial number to eink display
   * Add ability to display status on eink display